refactor(checkout): reorganize by feature
