--- libs/checkout/routing/src/guards/public_api.ts
@@ -1,2 +1,0 @@
-export { DaffCheckoutPlacedOrderGuard } from './placed-order/placed-order.guard';
-export * from './placed-order/placed-order-guard-redirect.token';
