+++ libs/checkout/routing/src/placed-order/public_api.ts
@@ -1,0 +1,2 @@
+export { DaffCheckoutPlacedOrderGuard } from './guard';
+export * from './guard-redirect.token';
