+++ libs/checkout/routing/src/public_api.ts
@@ -1,3 +1,3 @@
-export * from './guards/public_api';
+export * from './placed-order/public_api';
 
 export { DaffCheckoutRoutingModule } from './checkout-routing.module';
