--- libs/checkout/state/src/facades/public_api.ts
@@ -1,2 +1,0 @@
-export { DaffCheckoutPlacedOrderFacadeInterface } from './placed-order/placed-order-facade.interface';
-export { DaffCheckoutPlacedOrderFacade } from './placed-order/placed-order.facade';
