+++ libs/checkout/state/src/placed-order/public_api.ts
@@ -1,0 +1,6 @@
+export { DaffCheckoutPlacedOrderFacadeInterface } from './facade.interface';
+export { DaffCheckoutPlacedOrderFacade } from './facade';
+export {
+  DaffCheckoutPlacedOrderSelectors,
+  getCheckoutPlacedOrderSelectors,
+} from './selector';
