+++ libs/checkout/state/src/public_api.ts
@@ -1,5 +1,5 @@
 export * from './selectors/public_api';
-export * from './facades/public_api';
+export * from './placed-order/public_api';
 export * from './reducers/public_api';
 
 export { DaffCheckoutStateModule } from './checkout-state.module';
